% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBI.R
\name{EBI}
\alias{EBI}
\title{EBI}
\usage{
EBI(dataset, time = "time", status = "status", event = 1, threshold = 0.96)
}
\arguments{
\item{dataset}{The survival dataset with columns c(time, status, ...)}

\item{time}{The name of the column corresponding to patient follow-up time.}

\item{status}{The name of the column corresponding to patient status.}

\item{event}{A list of labels corresponding to an event in the status column.}
}
\value{
The result is the estimated censored bias index in the interval [0, 1].
A value over 0.5 suggests right-censoring information bias.
}
\description{
EBI estimates the bias index of a survival dataset attributed to censoring.
}
\examples{
aml_cbi <- EBI(survival::aml)

\dontrun{
 ovarian_cbi <- EBI(survival::ovarian) }
ovarian_cbi <- EBI(survival::ovarian, time = "futime", status = "fustat")

lung_cbi <- EBI(survival::lung, event = 2)
veteran_cbi <- EBI(survival::veteran)

}
\references{
Catalogue of bias collaboration. Bankhead CR, Spencer EA, Nunan D. Information bias.
 In: Sackett Catalogue Of Biases 2019.
 \href{https://catalogofbias.org/biases/information-bias}{Information bias}

 Barrajón E and Barrajón L (2020). Effect of right censoring bias on survival analysis.
 \href{https://arxiv.org/abs/2012.08649}{arXiv:2012.08649v1}
}
